Use socket instead of urequests when making HTTP GET towards QuestDB
When executing SQL and inserting data, the HTTP GET response from
QuestDB uses `Transfer-Encoding: chunked` which `urequests` doesn't
support.

This caused the code to always throw an exception. Catching the
exception is not the solution since urequests doesn't cleanup the
connection in that case and after few connections we get `ENOMEM`
error:

    OSError: [Errno 12] ENOMEM

We fix this issue by using raw sockets.